Fix code scanning alert 448 and 453: Possible loss of precision (#452)
* Potential fix for code scanning alert no. 448: Possible loss of precision


* #453 loss of precision fix

---------

+++ sqlnexus/Spinner.cs
@@ -320,7 +320,7 @@ namespace sqlnexus
         /// <returns>PointF object</returns>
         private PointF GetControlCenterPoint(Control _objControl)
         {
-            return new PointF(_objControl.Width / 2, _objControl.Height / 2);
+            return new PointF(_objControl.Width / 2f, _objControl.Height / 2f);
         }
 
         /// <summary>
@@ -375,7 +375,7 @@ namespace sqlnexus
         private double[] GetSpokesAngles(int _shtNumberSpoke)
         {
             double[] Angles = new double[_shtNumberSpoke];
-            double dblAngle = NUMBER_OF_DEGREES_CIRCLE / _shtNumberSpoke;
+            double dblAngle = (double)NUMBER_OF_DEGREES_CIRCLE / _shtNumberSpoke;
 
             for (int shtCounter = 0; shtCounter < _shtNumberSpoke; shtCounter++)
                 Angles[shtCounter] = (shtCounter == 0 ? dblAngle : Angles[shtCounter - 1] + dblAngle);
